Project completed

+++ EmployeeManagementSystem/src/main/java/com/ems/EmployeeManagementSystem/model/News.java
@@ -1,12 +1,12 @@
 package com.ems.EmployeeManagementSystem.model;
 
+import com.ems.EmployeeManagementSystem.model.enums.NewsTag;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.stereotype.Component;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -20,7 +20,7 @@ public class News {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(updatable = false)
-    private int newsId;
+    private long newsId;
 
     @Column(nullable = false)
     private String newsTitle;
