+++ EmployeeManagementSystem/src/main/java/com/ems/EmployeeManagementSystem/repository/NewsRepo.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NewsRepo extends JpaRepository<News, Integer> {
+public interface NewsRepo extends JpaRepository<News, Long> {
 }
